finished image filters, tags, and the download button
image filtering allows the user to apply blur and/or grayscale to an image and download the filtered image. Tags will allow the user to filter images for content they want to see. The user will have to input their own tags as the json does not contain descriptive tag data